requirements.txt
Flask==2.2.5
passlib==1.7.4
requests==2.28.1
beautifulsoup4==4.11.1
Werkzeug==2.2.3
APScheduler==3.10.4
python-dotenv==1.0.1
